add sortButton
For a sort to be shown, the sortButton will need to be clicked after one of the five sorting buttons has been clicked.

For testing's sake, I re-configured the code of the bubbleSort button so that one must click the sortButton after the bubbleSort button is clicked. This works as intended, and I plan to do the same for the other buttons. Moreover, I need to make it so that one cannot repeatedly click the sorting buttons.