add markButtonAsUnavailable, markButtonAsAvailable
These two functions set the text color of a button to red or its default (black) depending on which button has been clicked, and whether the sorting process is still going on. This was done so the user could easily keep track of which buttons are unable to be clicked when a given sorting algorithm is being complete.